per TA feedback, moved the API class and interface to an API folder nested inside data_access
