+++ src/data_access/API/GOStationApiClass.java
@@ -1,4 +1,4 @@
-package data_access;
+package data_access.API;
 
 import okhttp3.HttpUrl;
 import okhttp3.OkHttpClient;
@@ -12,7 +12,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 
-public class GOStationApiClass implements TrainApiInterface{
+public class GOStationApiClass implements TrainApiInterface {
 
     private final String PARTIAL_API_URL = "OpenDataAPI/api/V1";
     public GOStationApiClass () {
